tools/requirements.txt
supabase>=2.0.0
anthropic>=0.25.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
